pyproject.toml
[tool.poetry]
name = "instant-dashboard"
version = "0.1.0"
description = "InstantDashboard - Multi-agent data analytics assistant with natural language interface"
authors = ["InstantDashboard Team"]
license = "Apache License 2.0"
readme = "README.md"
package-mode = true


[tool.poetry.dependencies]
python = "^3.10"

# Core ADK dependencies
google-adk = "^1.0.0"
google-cloud-aiplatform = { extras = [
    "adk",
    "agent-engines",
], version = "^1.93.0" }

# Database and BigQuery dependencies
google-cloud-bigquery = "^3.28.0"
google-cloud-bigquery-storage = "^2.26.0"
db-dtypes = "^1.4.2"
sqlglot = "^26.10.1"

# Data processing dependencies
pandas = "^2.2.0"
numpy = "^1.26.0"

# Configuration and utilities
python-dotenv = "^1.0.1"
pydantic = { extras = ["email"], version = "^2.11.3" }
jsonschema = "^4.23.0"
requests = "^2.32.0"
tabulate = "^0.9.0"
email-validator = "^2.0.0"

# AI/ML dependencies
immutabledict = "^4.2.1"
absl-py = "^2.2.2"
regex = "^2024.11.6"

# FastAPI and web dependencies
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.34.0" }
python-multipart = "^0.0.18"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
pyjwt = { extras = ["crypto"], version = "^2.8.0" }
redis = "^5.2.1"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }

# NEW: Interactive Analytics Report Writer dependencies
# PostgreSQL and database management
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"

# WebSocket support for real-time collaboration
websockets = "^13.1"
python-socketio = "^5.11.4"

# Document management and export
python-docx = "^1.1.2"
weasyprint = "^63.0"
pypdf = "^5.1.0"

# Enhanced authentication and security
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"

# Additional utilities for report writer
jinja2 = "^3.1.4"
markdown = "^3.7"
bleach = "^6.2.0"


[tool.poetry.group.dev.dependencies]
google-cloud-aiplatform = { extras = [
    "adk",
    "agent-engines",
    "evaluation",
], version = "^1.93.0" }
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"


[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
#addopts = "-vv -s --pdb"
testpaths = ["tests/", "instant_dashboard/"]
log_level = "DEBUG"
log_cli = true
log_auto_indent = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = [
    "ignore::UserWarning",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
